+++ transport_tracker/authentication/templates/auth/base.html
@@ -8,10 +8,10 @@
     <meta name="description" content="Система отслеживания транспорта и прокладывания маршрутов">
     <meta name="keywords" content="транспорт, отслеживание, маршруты">
     <title>{% block title %}БелТранс{% endblock %}</title>
-    <link rel="stylesheet" href="{% static 'css/user/styles_main_base.css' %}">
-    <link rel="stylesheet" href="{% static 'css/user/styles_authentication.css' %}">
-    <link rel="stylesheet" href="{% static 'css/user/styles_contact.css' %}">
-    <link rel="stylesheet" href="{% static 'css/user/styles_about.css' %}">
+    <link rel="stylesheet" href="{% static 'css/styles_main_base.css' %}">
+    <link rel="stylesheet" href="{% static 'css/styles_authentication.css' %}">
+    <link rel="stylesheet" href="{% static 'css/styles_contact.css' %}">
+    <link rel="stylesheet" href="{% static 'css/styles_about.css' %}">
     <script defer src="{% static 'js/auth/scripts.js' %}"></script>
 </head>
 <body>
