+++ transport_tracker/authentication/templates/auth/register_page.html
@@ -25,6 +25,15 @@
                 <input type="email" class="form-control" id="email" name="email" placeholder="Введите электронную почту" required>
             </div>
             <div class="form-group">
+                <input type="text" class="form-control" id="first_name" name="first_name" placeholder="Введите имя" required>
+            </div>
+            <div class="form-group">
+                <input type="text" class="form-control" id="last_name" name="last_name" placeholder="Введите фамилию" required>
+            </div>
+            <div class="form-group">
+                <input type="text" class="form-control" id="contact_number" name="contact_number" placeholder="Введите контактный номер" required>
+            </div>
+            <div class="form-group">
                 <input type="password" class="form-control" id="password" name="password" placeholder="Введите пароль" required>
             </div>
             <div class="form-group">
