+++ transport_tracker/tracking/admin.py
@@ -1,5 +1,5 @@
 from django.contrib import admin
-from .models import Vehicle, Driver, DriverReview, VehicleType, Route, Trip  # Импорт всех моделей
+from .models import Vehicle, VehicleType, Route, Trip  # Импорт всех моделей
 
 # Админ-классы для транспортных средств
 class VehicleAdmin(admin.ModelAdmin):
@@ -12,16 +12,6 @@ class VehicleAdmin(admin.ModelAdmin):
     )
     search_fields = ('vehicle_number', 'driver__first_name', 'driver__last_name', 'vehicle_type__name')  # Поиск по типу транспорта
     list_filter = ('status', 'vehicle_type')  # Фильтрация по типу транспорта
-
-class DriverAdmin(admin.ModelAdmin):
-    list_display = ('first_name', 'last_name', 'contact_number', 'rating', 'review_count')
-    search_fields = ('first_name', 'last_name')
-    list_filter = ('rating',)
-
-class DriverReviewAdmin(admin.ModelAdmin):
-    list_display = ('driver', 'user', 'rating', 'created_at')
-    search_fields = ('driver__first_name', 'driver__last_name', 'user__username', 'rating')
-    list_filter = ('rating',)
 
 class VehicleTypeAdmin(admin.ModelAdmin):
     list_display = ('name', 'brand', 'max_capacity', 'year_of_manufacture', 'description')  # Показываем характеристики типа транспорта
@@ -39,8 +29,6 @@ class TripAdmin(admin.ModelAdmin):
 
 # Регистрация моделей в админке
 admin.site.register(Vehicle, VehicleAdmin)
-admin.site.register(Driver, DriverAdmin)
-admin.site.register(DriverReview, DriverReviewAdmin)
 admin.site.register(VehicleType, VehicleTypeAdmin)
 admin.site.register(Route, RouteAdmin)
 admin.site.register(Trip, TripAdmin)
