+++ transport_tracker/tracking/models.py
@@ -1,6 +1,7 @@
 from django.db import models
 from django.apps import apps
 from django.contrib.auth.models import User
+from driver.models import Driver
 
 # Модели для маршрутов и поездок (Map App)
 class Route(models.Model):
@@ -23,50 +24,6 @@ class Trip(models.Model):
 
     def __str__(self):
         return f"{self.vehicle} по маршруту {self.route} с {self.departure_time} до {self.arrival_time}"
-
-# Модели для водителей и транспортных средств (Tracking App)
-class Driver(models.Model):
-    first_name = models.CharField(max_length=50)
-    last_name = models.CharField(max_length=50)
-    contact_number = models.CharField(max_length=15)
-    rating = models.FloatField(default=0)
-    review_count = models.PositiveIntegerField(default=0)
-
-    def __str__(self):
-        return f"{self.first_name} {self.last_name}"
-
-    def update_rating(self):
-        reviews = self.reviews.all()
-        if reviews.exists():
-            total_rating = sum(review.rating for review in reviews)
-            self.rating = total_rating / reviews.count()
-            self.review_count = reviews.count()
-        else:
-            self.rating = 0
-            self.review_count = 0
-        self.save()
-
-    def can_take_trip(self, departure_time, arrival_time):
-        trips = Trip.objects.filter(vehicle__driver=self)
-        for trip in trips:
-            if (departure_time < trip.arrival_time and arrival_time > trip.departure_time):
-                return False
-        return True    
-
-
-class DriverReview(models.Model):
-    driver = models.ForeignKey(Driver, related_name='reviews', on_delete=models.CASCADE)
-    user = models.ForeignKey(User, related_name='driver_reviews', on_delete=models.CASCADE)
-    rating = models.IntegerField(choices=[(i, str(i)) for i in range(1, 6)])
-    comment = models.TextField()
-    created_at = models.DateTimeField(auto_now_add=True)
-
-    def save(self, *args, **kwargs):
-        super().save(*args, **kwargs)
-        self.driver.update_rating()
-
-    def __str__(self):
-        return f"Отзыв о {self.driver} от {self.user.username}: {self.rating} звезд"
 
 
 class VehicleType(models.Model):
